Changed Spinner to allow for Integer and Double values

+++ src/main/java/org/atdl4j/ui/javafx/widget/JavaFXSpinnerWidget.java
@@ -251,7 +251,7 @@ public class JavaFXSpinnerWidget extends AbstractJavaFXWidget<BigDecimal> {
                     if (control instanceof SingleSpinnerT) {
                         spinner.getValueFactory().setValue(Integer.parseInt(newValue));
                     } else if (control instanceof DoubleSpinnerT) {
-                        spinner.getValueFactory().setValue(Integer.parseInt(newValue));
+                        spinner.getValueFactory().setValue(Double.parseDouble(newValue));
                     }
                 }
             }
